requirements.txt
customtkinter
diffusers
huggingface-hub
numpy
Pillow
requests
tk
tokenizers
torch
tqdm
transformers
accelerate